[NuGet] Allow projects with no solution to get installed packages.
Xamarin.Insights requests the installed packages for a project when
a project is loaded when a solution is opened. At this point in time
the project does not have a parent solution. So a null reference was
thrown when PackageManagementProjectOperations.GetInstalledPackages
was called. Even if a project does not have a solution it is still
possible to find the installed packages from the packages.config file
or the project.json file. The PackageManagementProjectOperations
class now allows a project without a solution to find its installed
packages.