[NuGet] Add an async method for installing packages to public API.
Added InstallPackagesAsync to the PackageManagementProjectOperations
class that returns a Task. This is currently used by Xamarin.Insights
to add the Insights NuGet package and then waits for that to complete
before continuing.

Also fix an argument exception when Insights tries to use the
public API to get the installed packages when the project is still
being created from a project template. At this point the project may
not have a Name which causes an argument exception when using the
NuGet API's ProjectJsonPathUtilities.GetProjectConfigPath method. Now
if the project has now name then no package references are returned
for that project.