[NuGet] Ensure project.json file is reloaded after package action.
If the project.json file is open and a install, update or uninstall
occurs then the project.json file would not reflect the changes made
to the file. Now the project.json file is reloaded via the
FileService's FileChanged event. The code is directly calling the
BuildIntegrationProjectSystem's PostProcessAsync method. This is
called directly by the NuGetPackageManager for projects using a
packages.config but not if the project is using a project.json file
so as a workaround this is being called directly after an install,
update or uninstall. This PostProcessAsync method is used to
raise the FileService's FileChanged event for the project.json file.