[NuGet] Fix events not fired by public project operations API.
The PackageReferenceAdded and PackageReferenceRemoved events were not
fired by the PackageManagementProjectOperations class. The main impact
of this is if a NuGet package does not contain any assemblies then
removing it would not show an error for a Xamarin Component that
depended on it. Also installing that NuGet package back again would
not remove the error shown for the Xamarin Component. These events
are now fired after handling events generated by NuGet v3 for projects
that use a packages.config file. NuGet does not generate these events
for projects that use a project.json file so this problem still exists
in that case.