[NuGet] Support Roslyn analyzers being added or removed in the project
When a NuGet package is installed the file names in the NuGet package
are passed to the refactoring service. The refactoring service can
then an analyzer project item to the project. When a NuGet package is
removed the files are passed again to the refactoring service so it
can remove any analzyer project items in the project.

The above feature is currently disabled since the refactoring service
support for Roslyn analzyers is available on a separate feature
branch.