[Core] Fix task failure not reported for Runtime.RunInMainThread
If the Runtime.RunInMainThread (Func<Task> func) is called with a func
that uses await then the Task returned from RunInMainThread would not
report any failure of the task that it ran. When an await is used the
func itself does not throw an exception and instead returns a faulted
Task. The existing RunInMainThread then set a successful result on the
task it returns even though the task it ran failed.

Simple example below that reproduces the problem:

    Task task = Runtime.RunInMainThread (() => ThrowException ());
    task.Wait ();

async Task ThrowException ()
{
    await Task.Delay (1);
    throw new Exception ("error");
}

If there is no await in ThrowException then the original code in
RunInMainThread worked.