[NuGet] Fix packages not checked in Add Packages dialog after search.
If a NuGet package is checked in the Add Packages dialog and then
another search is run when the original NuGet package is displayed
again in the Add Packages dialog it was not checked. Now the NuGet
package is checked and the specific version that was selected is now
displayed in the version combo box.