[SourceEditor] Removed UnderlineError option.
This option doesn't really make any sense anymore and causes unwanted
side effects. It was used for disabeling compiler errors in older
versions.
The compiler errors are now shown as message bubbles which can be
disabled separately.
This here was now used for the on the fly errors and if that's
disabled the fix system wouldn't be usable anymore.