[NuGet] File conflict resoluton dialog now displayed.
If a file already exists and is modified when a NuGet package is
installed then now a dialog will be displayed asking the user if
they would like to overwrite the file or not. This dialog was not
being displayed after switching to NuGet v3 and now it is displayed
if necessary.