[NuGet] Fix null ref in Add Packages dialog.
On closing the Add Packages dialog a null reference exception could
occur if the package version information was returned after the
dialog was closed. The selected package view model's PropertyChanged
event handler was not being removed correctly and the package version
information was returned after the dialog was closed so the
selected package view model had already been set to null. This
resulted in a null reference exception:

Error loading package versions.
System.NullReferenceException: Object reference not set to an instance of an object
  at MonoDevelop.PackageManagement.AddPackagesDialog.SelectedPackageViewModelChanged (System.Object sender, System.ComponentModel.PropertyChangedEventArgs e) [0x00000] in main/src/addins/MonoDevelop.PackageManagement/MonoDevelop.PackageManagement.Gui/AddPackagesDialog.cs:723

Now the event handler is removed before another package is selected
and when the dialog is disposed.