[NuGet] Fixed files not removed when uninstalling NuGet package.
Uninstalling a NuGet package that has added files to a project, such
as jQuery, would not remove the files from the file system but was
removing them from the project file. Another side effect of this was
that if the files were added into a folder in the project, for
example Scripts, then the Solution window would still show this
folder with no files inside.

The problem was that the files were being removed from the project
but not deleted from the file system. Also NuGet 3 now requires
the project system to implement GetFiles and GetDirectories so
it can determine if a directory can be deleted.