[NuGet] Fix packages.config marked as deleted by Git.
The packages.config file was marked as deleted by Git if it had
one NuGet package referenced and an update, install or retarget
was run. This was because the file is deleted by NuGet and then
recreated. This bug was fixed previously in commit
1fd0463819aedab16bf36b578091a26e652dc04f but now different NuGet
package actions are used with NuGet v3 which broke this feature.