[NuGet] Fix retargeting not preserving local copy on references.
A retarget after the project's target framework has occurred will
uninstall and install the NuGet package. To preserve the reference
local copy value the NuGet addin needs to monitor both the uninstall
and the install of the NuGet package. What was happening was that
the local copy value was being preserved again during the install
but this is treated as a new install so the local copy is set back
to true. To prevent this the local copy preservation that is done
by the install action is disabled since it is already being handled
by the reinstall action.