[NuGet] Fixed no updates available message not shown in status area.
When updating a NuGet package if no updates are available then this
is now indicated in the status bar.

Note that for projects using a project.json files this works when
updating the NuGet packages for the project or solution but not
when updating an individual NuGet package. When a single NuGet
package is updated then NuGet always returns a single uninstall
action. This can be reproduced in Visual Studio 2015 when using the
PowerShell console and running Update-Package for a single NuGet
package which will show a message about uninstalling the NuGet
package.