[NuGet] Allow forcing a NuGet package uninstall.
If a NuGet package is not restored and an uninstall of that NuGet
package is attempted then a restore will be attempted. If the restore
succeeds then the NuGet package will be uninstalled. If the restore
fails then the user will be prompted asking if they want to force the
NuGet package to be removed. Forcing a NuGet package to be removed
will look for any references and import associated with the NuGet
package based on the packages install path. It will then remove the
NuGet package from the packages.config file. This feature is only
available if the project uses a packages.config file. A project that
uses a project.json file will not fail to uninstall since removing
the NuGet package will only modify the project.json file.