[NuGet] Fix solution specific NuGet.Config package sources not used.
Package sources can be defined for a solution in a NuGet.Config file
that is available alongside the solution or in any subdirectory above
the solution. This allows extra package sources to be defined for
the solution and will be displayed in the Add Packages dialog and
used for restoring and updating. The solution specific NuGet.Config
file was not being used and only the global NuGet.Config file store
under the user's profile ~/.config/NuGet/NuGet.Config was being used.