[NuGet] Use solution manager's settings when creating projects.
Ensure that whenever a solution is available its settings are used
when creating a NuGetProject. Previously the global NuGet.Config
settings were the only one used in a few places when a solution
settings was available. Using the global settings did not seem to
have any affect but just in case we use the solution manager settings
everywhere.