[NuGet] Fix license dialog shown for package already installed.
When installing the same NuGet package that required a license to
be accepted into a project when that NuGet package was already
installed into another project in the solution the license acceptance
dialog was displayed again. This would result in the same license
being accepted multiple times, once for each project. Now a check
is made to see if the NuGet package is installed into packages
directory for the solution to prevent the dialog being displayed
multiple times. This only works for projects that use packages.config
files. For project.json based projects the license acceptance dialog
is displayed for every project even if the NuGet package has been
installed into a project in the solution.