[NuGet] Fix package source changes not shown in Add Packages dialog.
Enabling or disabling a package source after a solution was opened
would not be reflected in the Add Packages dialog since the current
settings were cached for each solution. Now when the package sources
are changed in preferences the settings for each solution are reloaded
to ensure they are up to date so the correct package sources are
displayed in the Add Packages dialog.