[NuGet] Fixed credentials dialog not being displayed.
If a package source requires a username and a password but none have
been configured in the NuGet.Config file there was no dialog displayed
to ask for credentials from the user. With NuGet v3 the credential
provider that shows the credentials dialog needs to be connected to
the HttpHandlerResourceV3 class via its static methods.