[NuGet] Fixed package source errors not shown in Add Packages dialog.
Failures to connect to a package source were not shown in the Add
Packages dialog. Instead no packages were displayed if there was an
error. This is because the task being run to return the package
information does not directly report any errors. Instead error
information is logged by the MultiSourcePackageFeed. Now these logged
messages are displayed in the Add Packages dialog at the top.