[NuGet] Allow project templates with NuGet packages to work offline.
Previously with NuGet v2 the project templates used the official
nuget.org package source as the last package source when resolving
NuGet packages. With NuGet v2 even though the last package source
required internet access if the NuGet packages were available in a
local directory defined by an addin the project template could be
installed. With NuGet v3 this is no longer the case. NuGet v3 will
fail to install a NuGet package even if it is available in a local
directory if there is no internet access and the official nuget.org
package source is used. As a workaround an addin can indicate that
it does not need internet access for the NuGet packages in a project
template by setting a 'local' attribute to true for the NuGet package
reference.

<Package id="Xamarin.Forms" version="2.0.1.6495" local="true" />

If this set to true then the official nuget.org package source will
not be used and only local package sources will be used defined by
any addins using the extension path:

/MonoDevelop/Ide/ProjectTemplatePackageRepositories

This allows the project template to be installed without internet
access.