[Ide] Prepare text editor infrastructure for multiple caret/selection support.
Note that this is not breaking the public API in the first step.
Likeley is that some properties will get obsoleted. And some
operations may need to change and the API isn't yet completed. However
that could be postponed to the point when we've a control with multi
caret/selection support.