[NuGet] Support cancelling the current NuGet operation.
When the Stop button is clicked in the Package Console window the
current background NuGet operation will be cancelled. This is now
possible since NuGet v3 uses CancellationTokens throughout its API.
The progress monitor's CancellationToken is used with each NuGet
operation so clicking the Stop button will cancel that token which
will request that the current NuGet operation is cancelled.