[NuGet] Fixed versions not displayed in Add Packages dialog
After scrolling down the list of packages in the Add Packages dialog
so that another page of packages is requested from the package
source when a package was then selected the list of versions and
the dependency information would not be available. The problem was
that the CancellationToken was cancelled before the next page of
package data was loaded causing tasks associated with the packages
to be cancelled and not return any version or dependency data. Now
the token is cancelled only when new search terms are entered,
a different package source is selected or the dialog is closed.