[NuGet] Show recently installed packages in Add Packages dialog.
The behaviour is changed slightly from before. Now the recent packages
are filtered based on the selected package source. Previously the
same packages would be shown across all package sources. Now when
the search text is not an empty string no recent packages are shown.
Previously the recent packages would be shown but filtered based on
the search text.

Selecting a recent package will try to read the latest package
versions and metadata. Previously a recent package would not require
information from the package source.