[NuGet] Prevent package metadata being requested multiple times.
Every time a NuGet package was selected in the Add Packages dialog a
request was made to the package source for the package metadata even
if the data was already available. So selecting NUnit, then selecting
a different NuGet package, then selecting NUnit again would result
in the NUnit package metadata being requested again the second time it
was selected. This second request is now prevented. Note that if
different package source is selected or a new search term is entered
the request for the same NuGet package metadata will be made again
for the new search results.