[NuGet] Generate package events for project.json projects.
Using the actions when installing, updating or uninstalling a NuGet
package for a project that uses a project.json file now generates
package events. This allows a Xamarin Component to refresh its status
after a NuGet package is uninstalled or installed. Previously this
only worked with projects that used a packages.config file.