[NuGet] Fix UI hang for many package versions in Add Packages dialog.
When a NuGet package has a lot of package versions the Add Packages
dialog would become unresponsive for a small period of time when
populating the package version information in the combo box. One
example NuGet package is Paket which currently has 921 package
versions which would take just over one second to be added to the
combo box. During this one second the Add Packages dialog and the rest
of the UI was unresponsive as the UI thread was used to populate the
package version information.

Now if the package version information is loaded in sets of 100
versions at a time with a timer that runs periodically to load 100
more package versions until everything is populated. This allows the
UI to be responsive with the trade off that the package versions may
not immediately be available in the combo box.