[NuGet] Handle being unable to read the NuGet.Config file.
Report being unable to read the NuGet.Config file to the user when
opening a solution or the NuGet package sources page is opened in
Preferences.

Unfortunately with NuGet v3 even though the NuGet addin provides the
official NuGet package source it is not possible to use NuGet when
the NuGet.Config file cannot be read. Any access to the online package
source fails since it tries to read the NuGet.Config file which is
invalid.