[NuGet] Fix UI hang when searching for NuGet packages.
When searching for NuGet packages with the Add Packages dialog the
UI would hang occasionally. One problem was that calling
CancellationTokenSource.Cancel on the UI thread sometimes takes up to
a second on Mono. Another problem was that running the search would
run part of the logic on the UI thread which occasionally could take
a few seconds to run. Now both the cancellation and the search are
no longer run on the UI thread.