[NuGet] Fixed failing unit tests due to api change.
How the solution pad nodes display labels was changed in commit:

9197e1153e424ac964f50e6b444a7d393e459834

So instead of having one label there is a primary and a secondary
label. The secondary label is displayed in a different colour text.
This broke the unit tests since they were based on the original
api where there was no secondary label and the code in the NuGet
addin has been changed to use a secondary label.

Conflicts:
	main/external/fsharpbinding