[NuGet] Fixed jQuery summary badly formatted in Add Packages dialog
Newlines and empty space at the start and end of lines are now
removed in the summary that is displayed in the packages list in the
Add Packages dialog. This was working with NuGet v2 but a different
view model is used with NuGet v3 which did not fix the summary
text.