[NuGet] Fix no updates found when one package source fails.
When multiple package sources are enabled and one fails then the
Solution window would not show any package updates as being
available. This was because one package source failing would cause
the entire check for updates to fail. This is now handled and
update information is used from those package sources that are
not failing.