[NuGet] Fixed package readme.txt not being opened on update.
When updating an individual NuGet package or all packages in a
project the readme.txt of the updated NuGet package would not be
displayed. Note that this only works for projects that use a
packages.config file. Projects that use a project.json file do not
have an install action with the updated package version and do not
currently cause the readme.txt to be opened.