[NuGet] Improve handling of wildcard versions in Solution window.
A project using a project.json file can define a wildcard version:

"dependencies": {
    "Microsoft.NETCore.Platforms": "1.0.1-*"
}

Previously this was always shown as not installed since it the
version used was not known and not read from the project.lock.json
file. Also the version itself would be displayed as "1.0.1--" which
is neither the exact version nor the wildcard version.

Now NuGet packages that use wildcard version are always shown as
installed. There is no support currently for determining if the
NuGet package is missing. Also the version displayed in the context
menu when right clicking the package in the Solution window will
now show the wildcard version. Also in the Properties window the
wildcard version will be displayed as the NuGet package's version.