[NuGet] Fixed checked recent packages not being installed.
Fixed bug #41555 - Multiple packages can not be added simultaneously.
https://bugzilla.xamarin.com/show_bug.cgi?id=41555

After adding some NuGet packages with the Add Packages dialog,
then re-opening the Add Packages dialog and using the check box
for the recent packages at the top of the dialog then the recent
packages would not be installed. The problem was that the recent
packages still had the wrong parent so checking the package was
not causing it to be added to the list of checked packages.