[NuGet] Do not show status bar warnings about PowerShell scripts.
Fixed bug #41663 - Getting duplicate warning message in Package
Console Pad on adding the NuGet packages have PowerShell scripts.
https://bugzilla.xamarin.com/show_bug.cgi?id=41663

The bug refers to a duplicate message in the Package Console which
is not the real problem. The problem is that when a NuGet package
has PowerShell scripts the status bar should not show a warning.
Only the Package Console will have the warning.

This was changed before in 07953db18ed50cc547f92d7f95dfb6bc43973ff5
but when migrating to NuGet v3 the PowerShell message was changed
to be reported as a warning in the status bar. This has now been
changed so a warning is reported in the Package Console but not
in the status bar.