[NuGet] No longer show version at top right of Add Packages dialog.
Fixed bug #41669 - Showing wrong package version at the top right
corner of Add Package window when added old package version.
https://bugzilla.xamarin.com/show_bug.cgi?id=41669

When another version is selected in the version combo box the
version at the top right of the Add Packages dialog was not updated.
Since the version is now displayed in the version combo box there
is no need to show the version at the top right of the dialog so it
has been removed from there. So now there is no confusion about the
version that will be installed into the project.