[Core] Fix issue in multi-startup mode
When running in multi-startup mode, use the provided execution context only
for the first project, and use a default context for the other projects.

This is necessary because the debugger only supports starting one debugging
session.

Fixes bug #41756 - User is not able to run project in XS, when Multi Startup
option is selected