[NuGet] Fix delay for packages to appear in Add Packages dialog.
Fixed bug #41499 - Searching for a NuGet package takes five to ten
minutes to populate the results.
https://bugzilla.xamarin.com/show_bug.cgi?id=41499

When opening the Add Packages dialog sometimes it could take a while
(1 to 2 minutes) to display the list of NuGet packages. The problem
was that the images displayed in the Add Packages dialog for the
NuGet packages were loaded individually each using their own Task
which could use up many thread pool threads. The search itself uses a
Task and if this is run whilst the images are being loaded the task
can take a long time before a thread pool thread is made available to
start the search.

To fix this the image loading no longer uses tasks and instead uses
a single low priority background thread to load the images one at a
time.