[NuGet] Fix being unable install certain NuGet packages.
Fixed bug #41680 - Package are not downloaded automatically on
creation of android project.
https://bugzilla.xamarin.com/show_bug.cgi?id=41680

Certain NuGet packages when attempting to install them into a project
would fail with the error:

The UTC time represented when the offset is applied must be between
year 0 and 10,000.

This would occur if the current time zone is a positive offset from
UTC. This also only occurs for certain NuGet packages such as
XamarinComponent version 1.1.0.32.

A workaround has been added to NuGet itself to catch the
ArgumentOutOfRangeException that is thrown when the LastWriteTime
is retrieved from the zip archive entry. Previously NuGet would only
catch this exception when trying to write the file's last write time.