[NuGet] Remove more obsolete code.
Removed code that was loading the NuGet.Config file whenever a
solution was opened and was only used if the target framework of the
project was changed. Moved the loading of the NuGet.Config file into
the package compatibility instead. This allows a lot of unusued code
to be removed.