[Ide] Fix crash when opening project options
If a section is not added to the tree because none of their panels is visible
also remove it from the pages collection. Otherwise a client may request
to show that section, and MD will crash when trying to show it.

Fixes bug #41894 - Xamarin Studio crashes when trying to open "Options"
on an Android Bindings project.