[NuGet] Allow actions to be cancelled from the status bar.
All NuGet actions can now be cancelled from the status bar. The
NuGet action progress monitor enables the cancel button in the status
bar. This makes it easier to cancel the currently running NuGet
package action instead of having to open the Package Console and
click the cancel button there.