[Core] Remove framework check from TargetRuntime.ExecuteAssembly
ExecuteAssembly used to check if the framework selected for executing
an assembly is installed. This check is not really necessary, and in some
cases incorrect. On windows the framework check verifies that the redist
assemblies are available. This is necessary for building, but not for running.

Fixes bug #41658 - Cannot compile old projects. Starting a new solution
from scratch gives the same problem.

Fixes bug #37977 - C# consoleApp template doesn't build if .NET 4.5.1
is not installed (but 4.5 is)