[NuGet] Fix package compat messages not always shown in status bar
Fixed bug #41333 - On Windows, changing the Target Framework more than
one time, "Packages Retargeted Required" or "Packages are compatible"
message does not display in Status bar.
https://bugzilla.xamarin.com/show_bug.cgi?id=41333

Sometimes when changing the project's target framework for a
project the NuGet package compatibility checks status bar messages
would not be displayed and instead the status bar would just show
Project Saved. This is because the project save is asynchronous and
the package compatibility check is started before the save when the
target framework in the project is modified. If the Project Saved
message is displayed after the compatibility check is completed then
the status bar does not show the compatibility check message.

To fix this the package compatibility check waits for the project to
be saved before starting. This reduces the risk of the wrong status
bar message, Project Saved, being displayed. However it is still
possible that the Project Saved message will be shown, but unlikely, if
there is a long delay between the save completing and the Project
Saved message being sent to the status bar which may be after the
package compatibility check is completed. This does not seem to
happen in practice.