[NuGet] Fix incorrect button label in Add Packages dialog.
Fixed bug #42042 - When updating a Nuget package via
"Packages Dialog", the button is not changing to "Update Package"
https://bugzilla.xamarin.com/show_bug.cgi?id=42042

The text for the button was being created but the button's label was
not being updated. Also the plural string was being shown when no
packages were checked. This was broken in commit:

ed8070cd874497cc1303b82e527238c013a0e87b