[NuGet] Fix readme.txt being opened twice for same package
Fixed bug #42093 - When readme.txt from a package has already been
opened, adding same package to another project opens readme.txt again
https://bugzilla.xamarin.com/show_bug.cgi?id=42093

The readme.txt file was being opened again when the same NuGet
package was installed again into a different project in the same
solution.

Now a check is made to see if a NuGet package with the same version
is already installed into the packages directory before opening the
readme.txt file. Note that this only works for projects that use
a packages.config file and not for projects that use a project.json
file.