[NuGet] Always add .targets files to end of project.
When installing a NuGet package that has a .targets file the
Import element created was grouped with the existing Import elements.
This is OK most of the time however if there are other items in
the project added after the import then any build targets may fail
since these items are included after the import. One example is the
netfx-System.StringResources NuGet package which will not find any
resource files that are added to the project after the import.

Now .targets files are added as the last element in the project file.
This also makes the behaviour consistent with how NuGet works in
Visual Studio.