[NuGet] Fix All Sources not selected on re-opening Add Packages dialog
Fixed bug #42116 - Package source dropdown does not remember
"All Sources" option. It reset the package source and shows the first
package source instead of "All Sources"
https://bugzilla.xamarin.com/show_bug.cgi?id=42116

The All Sources was not being selected again in the Add Packages
dialog on re-opening the dialog when it was previously selected.
The problem was that the active package source matching was failing
since the All Sources has a different name, which is displayed,
compared with the package source name that is saved as the active
package source. This resulted in no package source match being found
and then the default behaviour was used where the first non-aggregate
package source is selected.