Make UnitTestGroup abstract
This class is never used directly, but always derived from. Because of
this, it makes more sense for this class to be abstract. This also
gives us the ability to mark some of the methods with a blank body as
abstract rather than cluttering the code up with unnecessary blank
methods which are misleading.