[Core] Fix policy being duplicated in Custom Policies dialog
Fixed #42146 - Each time a policy is edited, a policy duplicate is
shown with an appended ".mdpolicy" suffix
https://bugzilla.xamarin.com/show_bug.cgi?id=42146

With a custom policy available in the Custom Policies dialog, when
the OK button was clicked another policy file would be created.
On re-opening the IDE this would result in another copy of the
custom policy being shown in the Custom Policies dialog. Each time
this was done the policy copy would be created with another .mdpolicy
appended to the name.

The problem was that every time a policy was loaded the existing
name was replaced with the filename without its extension even
though the policy file contained a name. Policy files are saved with
the file extension .mdpolicy.xml so the name would have .mdpolicy
appended to it on starting the IDE again. Saving the policy again
when the OK button was pressed in the Custom Policies dialog would
then append .mdpolicy.xml to the policy name which already had
.mdpolicy as part of its name. This would cause more duplicates
every time the IDE was restarted.

f3cd5f3e686