[Refactoring] Fix analysis tasks not cleared on solution being closed
After analyzing the source the diagnostic errors and warnings are
added to TaskService and displayed in the Errors window. On closing
the solution and then opening another solution the status bar would
still show the warning and error icons for these diagnostics tasks.

The tasks created for the analysis errors and warnings are now
associated with the project or solution so they are removed when
the solution is closed.