[VCS] Optimize status querying
This refactors the code so both code paths use the same logic.

The problem of the old code was that for directory statuses, we'd query
the whole tree for conflicts and iterate all the index entries when
setting the info, thus doing an index refresh each time.

In comparison old and new case, we have these relative values:
128 OnGetDirectoryVersioninfoCalls with 222000ms
48 OnGetDirectoryVersionInfo calls with 50000ms

This optimizes by approximately 40%.