[C#] Improve speed of assemblybrowser and jumptodefinion for assemblybrowser.
Offloaded the parsing of the compilation to a background thread.

In the case of an assembly search, we now lookup only the types in the
current namespace, not all the types recursively found in child
namespaces, as we've already triggered a namespace search previously. If
no namespace matching the type that we're looking for is found, we end up
returning a LookupResult.Failure.

Removed double DFS search from the AssemblyBrowserWidget.