[Ide] Fix renaming a file containing ampersand shows another file
Fixed bug #35329 - Unable to rename a file with & in within the
resources directory
https://bugzilla.xamarin.com/show_bug.cgi?id=35329

In the solution pad when renaming a file that contained an ampersand
the label switched to another name. The problem was that the label
was being set to be the file name without the text being escaped.