[Ide] Optimize RecentFiles updates.
This modifies RecentFiles updates so that we do the following:
* Split Projects/Files into different events. This way we can guarantee
that for the WelcomePage, for example, we can subscribe only to project
changes that will cause a refresh of the items in the list.

* Don't trigger updates of recent items if nothing has changed
between the old and the new version. These updates can happen because
we send an IDE update event and also have the filesystemwatcher events
triggering updates for us.