[NuGet] Wait for restore to complete before building
Fixed bug #32155 - If I open a project and immediately try to build
it, the build will fail because package restoration is still occurring.
https://bugzilla.xamarin.com/show_bug.cgi?id=32155

If a NuGet package restore is in progress and a build is initiated the
build will wait until the package restore has finished before
starting. This prevents build errors when a build occurs whilst
NuGet package restore has not yet completed.