[Ide] Allow existing project files to be overwritten when adding files
Fixed bug #20204 - Adding existing items to a folder does not offer
overwrite option
https://bugzilla.xamarin.com/show_bug.cgi?id=20204

Adding existing files to a folder in a project would not allow the
existing files to be overwritten. The existing project files would
need to be deleted first. Now when adding an existing file to a folder
in a project where they already exist the user is prompted to
overwrite the existing file.