[NuGet] Wait for packages to be added before building new project
When a new project is created that has a lot of NuGet packages it
can take some time for the packages to be added. If a build is
started this can fail since the NuGet packages are not yet available.
Now when a new project is started and a build is started the build
will wait until the NuGet packages have been installed before
starting the compilation.