[NuGet] Improve build error message when NuGet package restore fails
When a build is waiting for NuGet packages to be restored and the
restore fails there was no useful information shown in the Errors
window for the build failure. Now when the NuGet package restore
fails and the build is waiting to start the restore error is
displayed as the build error in the Errors window. The build error
also shows that more details can be seen in the Package Console window.